requirements.txt
# python 3.12.0
torch
torchvision
scikit-learn
tqdm
pillow
pandas
matplotlib
seaborn
ipykernel
scikit-image
grad-cam
groundingdino-py